Renamed database schema dialect classes
