+++ src/Database/Dialect/SqlserverDialectTrait.php
@@ -22,8 +22,8 @@ use Cake\Database\Expression\UnaryExpression;
 use Cake\Database\ExpressionInterface;
 use Cake\Database\Query;
 use Cake\Database\QueryCompiler;
-use Cake\Database\Schema\BaseSchema;
-use Cake\Database\Schema\SqlserverSchema;
+use Cake\Database\Schema\SchemaDialect;
+use Cake\Database\Schema\SqlserverSchemaDialect;
 use Cake\Database\SqlDialectTrait;
 use Cake\Database\SqlserverCompiler;
 use Cake\Database\ValueBinder;
@@ -330,11 +330,11 @@ trait SqlserverDialectTrait
      * Used by Cake\Schema package to reflect schema and
      * generate schema.
      *
-     * @return \Cake\Database\Schema\BaseSchema
+     * @return \Cake\Database\Schema\SchemaDialect
      */
-    public function schemaDialect(): BaseSchema
+    public function schemaDialect(): SchemaDialect
     {
-        return new SqlserverSchema($this);
+        return new SqlserverSchemaDialect($this);
     }
 
     /**
